requirements.txt
Django==3.2.14
djangorestframework==3.13.1
django-allauth==0.50.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.7.2
django-filter==21.1
python-dotenv==1.0.1